+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     - Errors occurring while initially loading databases
       (or when calling `Geolix.reload_databases/0`) are now sent to
       `Logger.error` ([#16](https://github.com/elixir-geolix/geolix/issues/16))
+    - "GeoLite2-ASN" databases return structs instead of plain maps
     - When an error occurs during database load the database will be
       excluded from the lookup process